TextReplace#2: File Structure changed and modules renamed.
